+++ src/main.js
@@ -1,4 +1,4 @@
-import { createRepositoriesService } from "./factory";
+import { createRepositoriesService } from "./factory/index.js";
 
 const repositoriesService = createRepositoriesService();
 repositoriesService.main();
